Refactor: extract util functions to own file
